+++ src/components/defaultSettingsModal/CdmApiTokenToggle.js
@@ -1,22 +1,13 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
 import List from "@material-ui/core/List";
 import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemSecondaryAction from "@material-ui/core/ListItemSecondaryAction";
 import ListItemText from "@material-ui/core/ListItemText";
-import ListSubheader from "@material-ui/core/ListSubheader";
 import Switch from "@material-ui/core/Switch";
-import CodeIcon from "@material-ui/icons/Code";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    width: "100%",
-  },
-}));
+import AssignmentIndIcon from "@material-ui/icons/AssignmentInd";
 
 export default function SwitchListSecondary() {
-  const classes = useStyles();
   const [checked, setChecked] = React.useState(["wifi"]);
 
   const handleToggle = (value) => () => {
@@ -36,11 +27,11 @@ export default function SwitchListSecondary() {
     <List>
       <ListItem>
         <ListItemIcon>
-          <CodeIcon />
+          <AssignmentIndIcon />
         </ListItemIcon>
         <ListItemText
           id="switch-list-label-wifi"
-          primary="Polaris Development Mode"
+          primary="CDM API Token Authentication"
         />
         <ListItemSecondaryAction>
           <Switch
