+++ src/components/defaultSettingsModal/PolarisDevelopmentModeToggle.js
@@ -1,22 +1,13 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
 import List from "@material-ui/core/List";
 import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemSecondaryAction from "@material-ui/core/ListItemSecondaryAction";
 import ListItemText from "@material-ui/core/ListItemText";
-import ListSubheader from "@material-ui/core/ListSubheader";
 import Switch from "@material-ui/core/Switch";
-import AssignmentIndIcon from "@material-ui/icons/AssignmentInd";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    width: "100%",
-  },
-}));
+import CodeIcon from "@material-ui/icons/Code";
 
 export default function SwitchListSecondary() {
-  const classes = useStyles();
   const [checked, setChecked] = React.useState(["wifi"]);
 
   const handleToggle = (value) => () => {
@@ -36,11 +27,11 @@ export default function SwitchListSecondary() {
     <List>
       <ListItem>
         <ListItemIcon>
-          <AssignmentIndIcon />
+          <CodeIcon />
         </ListItemIcon>
         <ListItemText
           id="switch-list-label-wifi"
-          primary="CDM API Token Authentication"
+          primary="Polaris Development Mode"
         />
         <ListItemSecondaryAction>
           <Switch
